Considering issues with times being in unix leap time.
I put in a correction to the atomic_difference which fixes that, but
I am wondering if I should always consider system_time to be unix time
or maybe it should be unix leap time (if leap seconds are defined). The
weird thing is that days are exactly every 24 hours in unix leap time,
meaning that the unix leap day is 27 seconds before the unix day, but
daylight savings shifts are at the same time, meaning that daylight
shifts start at 2:00:27 am in unix leap time. Also, most time zone files
ignore the existance of leap seconds.

Maybe I should have a leaptimestamp? I am not exactly sure how to
best proceed.